+++ adapter/memQuota/BUILD
@@ -7,10 +7,10 @@ go_library(
     srcs = [
         "memQuota.go",
         "rollingWindow.go",
-        "util.go",
     ],
     deps = [
         "//adapter/memQuota/config:go_default_library",
+        "//adapter/memQuota/util:go_default_library",
         "//pkg/adapter:go_default_library",
         "//pkg/pool:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
